+++ src/Universum/Lifted/File.hs
@@ -6,9 +6,7 @@
 
 module Universum.Lifted.File
        ( appendFile
-       , getContents
        , getLine
-       , interact
        , openFile
        , readFile
        , writeFile
@@ -20,8 +18,6 @@ import Prelude (FilePath)
 import System.IO (Handle, IOMode)
 
 import qualified Data.Text.IO as XIO
-import qualified Data.Text.Lazy as L (Text)
-import qualified Data.Text.Lazy.IO as LIO (getContents, interact)
 import qualified System.IO as XIO (openFile)
 
 ----------------------------------------------------------------------------
@@ -33,22 +29,10 @@ appendFile :: MonadIO m => FilePath -> Text -> m ()
 appendFile a b = liftIO (XIO.appendFile a b)
 {-# INLINE appendFile #-}
 
--- | Lifted version of 'Data.Text.getContents'.
-getContents :: MonadIO m => m L.Text
-getContents = liftIO LIO.getContents
-{-# INLINE getContents #-}
-{-# DEPRECATED getContents "This function will be removed in a future version of this package, use `liftIO` with `Data.Text.Lazy.IO.getContents` from `text` instead." #-}
-
 -- | Lifted version of 'Data.Text.getLine'.
 getLine :: MonadIO m => m Text
 getLine = liftIO XIO.getLine
 {-# INLINE getLine #-}
-
--- | Lifted version of 'Data.Text.interact'.
-interact :: MonadIO m => (L.Text -> L.Text) -> m ()
-interact a = liftIO (LIO.interact a)
-{-# INLINE interact #-}
-{-# DEPRECATED interact "This function will be removed in a future version of this package, use `liftIO` with `Data.Text.Lazy.IO.interact` from `text` instead." #-}
 
 -- | Lifted version of 'Data.Text.readFile'.
 readFile :: MonadIO m => FilePath -> m Text